Add docker service and publish action
This adds the ability to run this logger from docker, which
makes running it in the background easier, on many platforms.
For further convenience, we add a GitHub action that auto
builds this container and publishes it to the GitHub container
repository associated with this repo.

We enable qemu and buildx, to allow running this image on
Raspberry Pi (arm64).

Example run line:
docker run --rm -it ghcr.io/ratgdo/remote-logging 192.168.1.123